Allow absent legacy folders

+++ filebrowser/rootfs/etc/cont-init.d/20-folders.sh
@@ -18,7 +18,9 @@ fi
 
 # Clean symlinks
 find /config -maxdepth 1 -type l -delete
-find /homeassistant/addons_config -maxdepth 1 -type l -delete
+if [ -d /homeassistant/addons_config ]; then
+    find /homeassistant/addons_config -maxdepth 1 -type l -delete
+fi
 
 # Remove erroneous folders
 if [ -d /homeassistant ]; then
@@ -30,7 +32,11 @@ if [ -d /homeassistant ]; then
     fi
 fi
 
-# Create symlinks
-ln -s /homeassistant/addons_config /config
-ln -s /homeassistant/addons_autoscripts /config
-find /addon_configs/ -maxdepth 1 -mindepth 1 -type d -not -name "*filebrowser*" -exec ln -s {} /config/addons_config/ \;
+# Create symlinks with legacy folders
+if [ -d /homeassistant/addons_config ]; then
+    ln -s /homeassistant/addons_config /config
+    find /addon_configs/ -maxdepth 1 -mindepth 1 -type d -not -name "*filebrowser*" -exec ln -s {} /config/addons_config/ \;
+fi
+if [ -d /homeassistant/addons_autoscripts ]; then
+    ln -s /homeassistant/addons_autoscripts /config
+fi
